move finding of the ffprobe executable to it’s own class
