+++ lib/ffprober.rb
@@ -8,26 +8,10 @@ require_relative 'ffprober/video_stream'
 require_relative 'ffprober/subtitle_stream'
 require_relative 'ffprober/chapter'
 require_relative 'ffprober/ffprobe_version'
+require_relative 'ffprober/ffmpeg/finder'
 require 'json'
 
 module Ffprober
-  def self.path
-    @path ||= begin
-      path = ENV['PATH'].split(File::PATH_SEPARATOR).find do |path_to_check|
-        File.executable?(File.join(path_to_check, executable_name))
-      end
-
-      path && File.expand_path(executable_name, path)
-    end
-  end
-
-  def self.executable_name
-    @executable_name ||= self.windows? ? 'ffprobe.exe' : 'ffprobe'
-  end
-
-  def self.windows?
-    !!(RUBY_PLATFORM =~ /(mingw|mswin)/)
-  end
 
   class InvalidInputFileError < ::StandardError; end
 end
