+++ spec/ffprober/ffmpeg/finder_spec.rb
@@ -1,11 +1,11 @@
 # encoding: utf-8
 require 'spec_helper'
 
-describe Ffprober do
+describe Ffprober::Ffmpeg::Finder do
 
   describe 'if no ffprobe is found' do
     it 'should raise a exception if there is no ffmpeg' do
-      allow(Ffprober).to receive(:path).and_return('nonexistant')
+      allow(described_class).to receive(:path).and_return('nonexistant')
 
       expect do
         Ffprober::Parser.from_file('spec/assets/301-extracting_a_ruby_gem.m4v')
